backend/requirements.txt
flask
flask-cors
confluent-kafka
python-dotenv
opencv-python
numpy